completed logic for add, update, and delete bmw models based on user permission

+++ controllers/bmw.controller.js
@@ -4,7 +4,7 @@ export const createBmw = async (req, res, next) => {
     try{
         const body = {...req.body, createdBy: req.user._id};
 
-        if(req.user.isAdmin && body.isOfficial === true){
+        if(!req.user.isAdmin && body.isOfficial === true){
             return res.status(403).json({success: false, error: 'Only admins can mark official models'});
         }
 
@@ -20,6 +20,9 @@ export const createBmw = async (req, res, next) => {
 
 export const updateBmw = async (req, res, next) => {
     try{
+        if(!req.user.isAdmin && req.body.isOfficial === true){
+            return res.status(403).json({success: false, error: 'Only admins can update official models'});
+        }
         const updated = await Bmw.findByIdAndUpdate(req.params.id, req.body,{new:true, runValidators:true});
         if(!updated){
             const error = new Error('Bmw not found');
@@ -36,6 +39,9 @@ export const updateBmw = async (req, res, next) => {
 
 export const deleteBmw = async (req, res, next) => {
     try{
+        if(!req.user.isAdmin && req.body.isOfficial === true){
+            return res.status(403).json({success: false, error: 'Only admins can delete official models'});
+        }
         const deleted = await Bmw.findByIdAndDelete(req.params.id);
         if(!deleted){
             const error = new Error('Bmw not found');
